+++ packages/python/src/axonpush/_internal/api/models/__init__.py
@@ -94,7 +94,6 @@ from .create_experiment_dto_model_configuration import CreateExperimentDtoModelC
 from .create_export_destination_dto import CreateExportDestinationDto
 from .create_export_destination_dto_headers import CreateExportDestinationDtoHeaders
 from .create_invitation_dto import CreateInvitationDto
-from .create_iot_token_dto import CreateIotTokenDto
 from .create_online_rule_dto import CreateOnlineRuleDto
 from .create_organization_dto import CreateOrganizationDto
 from .create_prompt_dto import CreatePromptDto
@@ -183,9 +182,6 @@ from .intelligence_job_response_dto import IntelligenceJobResponseDto
 from .intelligence_job_response_dto_clusters_item import IntelligenceJobResponseDtoClustersItem
 from .intelligence_job_status import IntelligenceJobStatus
 from .invitation_response_dto import InvitationResponseDto
-from .iot_credentials_response_dto import IotCredentialsResponseDto
-from .iot_token_create_response_dto import IotTokenCreateResponseDto
-from .iot_token_response_dto import IotTokenResponseDto
 from .issue_occurrence_response_dto import IssueOccurrenceResponseDto
 from .issue_occurrence_response_dto_evidence import IssueOccurrenceResponseDtoEvidence
 from .issue_response_dto import IssueResponseDto
@@ -423,7 +419,6 @@ __all__ = (
     "CreateExportDestinationDto",
     "CreateExportDestinationDtoHeaders",
     "CreateInvitationDto",
-    "CreateIotTokenDto",
     "CreateOnlineRuleDto",
     "CreateOrganizationDto",
     "CreatePromptDto",
@@ -512,9 +507,6 @@ __all__ = (
     "IntelligenceJobResponseDtoClustersItem",
     "IntelligenceJobStatus",
     "InvitationResponseDto",
-    "IotCredentialsResponseDto",
-    "IotTokenCreateResponseDto",
-    "IotTokenResponseDto",
     "IssueOccurrenceResponseDto",
     "IssueOccurrenceResponseDtoEvidence",
     "IssueResponseDto",